color picker is done for iPhone and iPad

+++ Makestagram/NEOPickerSources/NEOColorPickerFavoritesViewController.m
@@ -71,9 +71,9 @@
     
     NSOrderedSet *colors = [NEOColorPickerFavoritesManager instance].favoriteColors;
     UIColor *pattern = [UIColor colorWithPatternImage:[UIImage imageNamed:@"colorPicker.bundle/color-picker-checkered"]];
-    int count = [colors count];
-    for (int i = 0; i < count; i++) {
-        int page = i / 24;
+    NSUInteger count = [colors count];
+    for (NSUInteger i = 0; i < count; i++) {
+        NSUInteger page = i / 24;
         int x = i % 24;
         int column = x % 4;
         int row = x / 4;
@@ -93,7 +93,7 @@
         [self setupShadow:layer];
         [self.scrollView.layer addSublayer:layer];
     }
-    int pages = (count - 1) / 24 + 1;
+    NSUInteger pages = (count - 1) / 24 + 1;
     
     self.scrollView.contentSize = CGSizeMake(pages * 320, 296);
     self.scrollView.delegate = self;
